Safari-611.1.10/LayoutTests/platform/wk2/loader/go-back-cached-main-resource-expected.txt

go-back-cached-main-resource.html - didFinishLoading
resources/first-page.html - willSendRequest <NSURLRequest URL resources/first-page.html, main document URL resources/first-page.html, http method GET> redirectResponse (null)
resources/first-page.html - didReceiveResponse <NSURLResponse resources/first-page.html, http status code 0>
resources/first-page.html - didFinishLoading
resources/other-page.html - willSendRequest <NSURLRequest URL resources/other-page.html, main document URL resources/other-page.html, http method GET> redirectResponse (null)
resources/other-page.html - didReceiveResponse <NSURLResponse resources/other-page.html, http status code 0>
resources/other-page.html - didFinishLoading
resources/first-page.html - willSendRequest <NSURLRequest URL resources/first-page.html, main document URL resources/first-page.html, http method GET> redirectResponse (null)
resources/first-page.html - didReceiveResponse <NSURLResponse resources/first-page.html, http status code 0>
resources/first-page.html - didFinishLoading
resources/other-page.html - willSendRequest <NSURLRequest URL resources/other-page.html, main document URL resources/other-page.html, http method GET> redirectResponse (null)
resources/other-page.html - didReceiveResponse <NSURLResponse resources/other-page.html, http status code 0>
resources/other-page.html - didFinishLoading
resources/first-page.html - willSendRequest <NSURLRequest URL resources/first-page.html, main document URL resources/first-page.html, http method GET> redirectResponse (null)
resources/first-page.html - didReceiveResponse <NSURLResponse resources/first-page.html, http status code 0>
resources/first-page.html - didFinishLoading
This test check the following situation:

First you navigate to a page (first-page.html).
Then you go to another page (other-page.html).
You repeat previous steps again (going to first-page.html and then to other-page.html).
Finally you click back.
The problem was that the resource load callbacks when going back to the cached main resource were not being called.

See bug #112418.
